Moved UI components to UI folder and renamed NetworkSelector to WalletProviderSelector
