+++ package.json
@@ -14,7 +14,7 @@
     "deploy": "rm -rf out && npm run build:static",
     "deploy:legacy": "npm run deploy:prebuilt && gh-pages --dotfiles --dist out",
     "deploy:prebuilt": "rm -rf out && npm run build:static && touch out/.nojekyll && npm run cname",
-    "integration": "testcafe chrome **/**/*.spec.js",
+    "integration": "testcafe chrome src/**/*.spec.js",
     "integration:headless": "npm run build:static && testcafe chrome:headless src/**/*.spec.js --app \"npm run serve-static\"",
     "integration:watch": "nodemon --exec \"npm run integration\" --watch src"
   },
