+++ src/components/UI/HashColor.js
