+++ src/components/UI/PageLoader/style.scss
