+++ src/components/UI/Panel/style.scss
