+++ src/components/WalletProviderSelector/style.scss
