+++ src/components/WalletProviderSelectorPage.js
@@ -1,14 +1,14 @@
 import React from "react";
-import Panel from "./Panel";
-import NetworkSelector from "./NetworkSelector";
+import Panel from "./UI/Panel";
+import WalletProviderSelector from "./WalletProviderSelector";
 
-const NetworkSelectorPage = () => (
+const WalletProviderSelectorPage = () => (
   <Panel id="network-selector" style={{ textAlign: "center" }}>
     <img src="/static/images/logo.svg" style={{ maxWidth: 300 }} />
     <h1>Welcome to OpenCerts Admin Portal</h1>
     <p>Please select a Wallet Network to begin.</p>
-    <NetworkSelector />
+    <WalletProviderSelector />
   </Panel>
 );
 
-export default NetworkSelectorPage;
+export default WalletProviderSelectorPage;
